Start work on plugin system
Explaination of PluginHost and the WIP GUI system:

I really didn't want to have to make two separate libs across two different frameworks to enable plugin creation for both the GUI level and server level, so all the plugin loading is going to be done at the server level (this allows SimpleServer to still work even if I decide to port the GUI to Java to enable cross-platform support). The server will constantly send "ControlUpdate" requests to a listener that the GUI creates containing the JSON version of the GUI controls. A new one of these requests get sent every time a plugin trys to edit a GUI control. The server will also create a request every 50ms, and the GUI will send back a list of events (such as button clicks) that have occured since the last request.